long, not int

+++ core/src/main/scala/org/broadinstitute/dsde/rawls/Boot.scala
@@ -130,7 +130,7 @@ object Boot extends IOApp with LazyLogging {
         gcsConfig.getString("subEmail"),
         gcsConfig.getString("pathToPem"),
         gcsConfig.getString("appsDomain"),
-        dmConfig.getInt("orgID"),gcsConfig.getString("groupsPrefix"),
+        dmConfig.getLong("orgID"),gcsConfig.getString("groupsPrefix"),
         gcsConfig.getString("appName"),
         gcsConfig.getInt("deletedBucketCheckSeconds"),
         serviceProject,
